Posts一覧をVue.jsでコンポーネント化
