+++ app/controllers/application_controller.rb
@@ -1,6 +1,11 @@
 class ApplicationController < ActionController::Base
   before_action :basic_auth
 
+  def shared_data
+    @shared_data ||= {}
+  end
+  helper_method :shared_data
+
 
   private
