+++ app/javascript/test.js
@@ -3,6 +3,7 @@ function test() {
 
   let parseStr = qs.parse('station=tama-plaza&num=16');
   console.log(parseStr)
+  console.log(sharedData)
 }
 
 document.addEventListener('DOMContentLoaded', test)
