+++ app/views/layouts/application.html.erb
@@ -12,5 +12,6 @@
 
   <body>
     <%= yield %>
+    <%= render 'shared/shared_data' %>
   </body>
 </html>
