+++ app/views/posts/index.html.erb
@@ -5,14 +5,3 @@
 <% end %>
 
 <div id="app"></div>
-
-<% @posts.each do |post| %>
-  <div class="post" data-id=<%= post.id %> data-check=<%= post.checked %>>
-    <div class="post-date">
-      投稿日時：<%= post.created_at %>
-    </div>
-    <div class="post-content">
-      <%= post.content %>
-    </div>
-  </div>
-<% end %>
