+++ app/views/posts/index.json.jbuilder
@@ -1,0 +1,3 @@
+json.posts do
+  json.array! @posts, :id, :content, :checked, :created_at, :updated_at
+end
