+++ app/views/shared/_shared_data.html.erb
@@ -1,0 +1,3 @@
+<script>
+let sharedData = <%= sanitize shared_data.to_json %>;
+</script>
